medicines service done

+++ src/main/java/com/example/Meds/repository/MedicineRepository.java
@@ -3,5 +3,5 @@ package com.example.Meds.repository;
 import com.example.Meds.entity.Medicine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicineRepository extends JpaRepository<Medicine, Integer> {
+public interface MedicineRepository extends JpaRepository<Medicine, Long> {
 }
